refactor(features/cards): create pages, move to @api
issues #66, #65
