+++ src/features/cards/index.js
@@ -1,11 +1,10 @@
-import "./model"
+// @flow
 
-import { type Card } from "./types"
-
-export { $registry as $cardsRegistry } from "./model/registry.store"
-export { cardsToObject } from "./model/registry.model"
-export { CardItem, CardsList, SkeletonList } from "./organisms"
+export {
+  $registry as $cardsRegistry,
+  cardsToObject,
+} from "./model/registry.store"
+export { usefulMarkClicked } from "./model/registry.events"
+export { CardItem, CardsList, SkeletonList, CardSkeleton } from "./organisms"
 export { CardsCommonTemplate } from "./templates/common"
-export { cardsRoutes } from "./routes"
-
-export { Card }
+export { TitleInput } from "./atoms/title-input"
