--- src/features/cards/model/index.js
@@ -1,1 +1,0 @@
-import "./registry.model"
