--- src/features/cards/routes.js
@@ -1,27 +1,0 @@
-import { CardCreatePage } from "./pages/create"
-import { CardsHomePage } from "./pages/home"
-import { CardViewPage } from "./pages/view"
-import { CardEditPage } from "./pages/edit"
-
-export const cardsRoutes = () => [
-  {
-    path: "/",
-    exact: true,
-    component: CardsHomePage,
-  },
-  {
-    path: "/new/card",
-    exact: true,
-    component: CardCreatePage,
-  },
-  {
-    path: "/open/:cardId",
-    exact: true,
-    component: CardViewPage,
-  },
-  {
-    path: "/edit/:cardId",
-    exact: true,
-    component: CardEditPage,
-  },
-]
