+++ src/pages/create/model.js
@@ -1,3 +1,4 @@
+// @flow
 import Plain from "slate-plain-serializer"
 import {
   createEvent,
@@ -5,20 +6,21 @@ import {
   createStore,
   createStoreObject,
 } from "effector"
-import { createFetching } from "@lib/fetching"
-import { history } from "@lib/routing"
-import { cardsApi } from "../api"
 
-export const titleChanged = createEvent()
-export const contentChanged = createEvent()
-export const formSubmitted = createEvent()
-export const pageUnmounted = createEvent()
+import { createFetching, type Fetching } from "@lib/fetching"
+import { history } from "@lib/routing"
+import { cardsApi, type Card } from "@api/cards"
+
+export const titleChanged = createEvent<SyntheticEvent<HTMLInputElement>>()
+export const contentChanged = createEvent<*>()
+export const formSubmitted = createEvent<*>()
+export const pageUnmounted = createEvent<void>()
 
 const cardCreate = createEffect()
-export const cardCreateFetching = createFetching(cardCreate)
+export const cardCreateFetching: Fetching<Card, *> = createFetching(cardCreate)
 
-export const $title = createStore("")
-export const $content = createStore(
+export const $title = createStore<string>("")
+export const $content = createStore<{}>(
   Plain.deserialize("Start typing here...").toJS(),
 )
 const $form = createStoreObject({
@@ -26,7 +28,7 @@ const $form = createStoreObject({
   content: $content,
 })
 
-const trimEvent = (event) => event.target.value
+const trimEvent = (event) => event.currentTarget.value
 
 $title.on(titleChanged.map(trimEvent), (_, title) => title)
 $content.on(contentChanged, (_, content) => content)
@@ -36,7 +38,7 @@ formSubmitted.watch(() => {
   cardCreate($form.getState())
 })
 
-cardCreate.use((form) => cardsApi.create(form))
+cardCreate.use(cardsApi.create)
 
 cardCreate.done.watch(({ result: card }) => {
   history.push(`/open/${card.id}`)
