+++ src/pages/settings/page.js
@@ -1,0 +1,4 @@
+// @flow
+import * as React from "react"
+
+export const SettingsPage = () => <div />
