+++ src/pages/settings/store.js
@@ -1,0 +1,1 @@
+// @flow
