+++ src/pages/view/model.js
@@ -1,14 +1,12 @@
 // @flow
 import { createEffect, createStore, createEvent, combine } from "effector"
-import type { Event, Effect, Store } from "effector"
+import type { Effect, Store } from "effector"
 
 import { createFetching, type Fetching } from "@lib/fetching"
-import { cardsApi } from "../api"
-import type { Card } from "../types"
-import { $registry } from "./registry.store"
-import { cardsToObject } from "./registry.model"
+import { cardsApi, type Card } from "@api/cards"
+import { $cardsRegistry, cardsToObject } from "@features/cards"
 
-export const pageUnloaded: Event<void> = createEvent()
+export const pageUnloaded = createEvent<void>()
 
 export const cardLoading: Effect<
   { cardId: number },
@@ -23,7 +21,7 @@ export const cardFetching: Fetching<*, void> = createFetching(
 const $cardId = createStore(-1)
 
 export const $card: Store<?Card> = combine(
-  $registry,
+  $cardsRegistry,
   $cardId,
   (registry, cardId) => registry[cardId],
 )
@@ -34,7 +32,7 @@ $cardId.on(cardLoading, (_, { cardId }) => cardId)
 $cardId.on(cardLoading.done, (_, { result }) => result.card.id)
 $cardId.reset(pageUnloaded)
 
-$registry.on(cardLoading.done, (registry, { result }) => {
+$cardsRegistry.on(cardLoading.done, (registry, { result }) => {
   return {
     ...registry,
     ...cardsToObject([result.card]),
